Updating unittests for pipelines and templates
