+++ test/templates/dependencies_2.jst
@@ -1,0 +1,10 @@
+# Basic template tests hello world in two steps
+# Expected stdout: "Hello, world!" 
+- name: hello
+  output: hello.txt
+  cmd: printf 'Hello, '
+
+- name: world
+  input: hello.txt
+  cmd: printf 'world!\n'
+
