+++ test/templates/dependencies_3.jst
@@ -1,0 +1,22 @@
+# This template tests dynamic features with the exec directive. 
+# During a run, add_tasks will add a new to the workflow. The
+# last task will wait for the new task to complete via the 
+# after-re directive.
+# Expected stdout: "Hello, world! All done!"
+- name: start
+  cmd: printf 'Hello, ' 
+
+
+- name: add_tasks
+  after: start
+  exec: |
+    runner.workflow.new_task(
+        name='dynamic_task',
+        cmd="printf 'world! '"
+    )
+
+
+- name: after_adding_tasks
+  after-re: dynamic_task|add_tasks
+  cmd: printf 'All done!\n'
+
