--- test/templates/hello_world.jst
@@ -1,2 +1,0 @@
-- name: hello_world
-  cmd: null
