+++ test/templates/helloworld_1.jst
@@ -1,0 +1,5 @@
+# Basic workflow tests hello world example
+# Expected stdout: Hello, world! 
+- name: hello_world
+  cmd: echo 'Hello, world!'
+
