+++ test/templates/inheritance_1.jst
@@ -1,0 +1,7 @@
+# This template tests inheritance with the include statement
+# Expected stdout: "Hello, world!"
+{% include 'helloworld_1.jst' %}
+
+- name: task1
+  cmd: null
+
