--- test/templates/null.jst
@@ -1,8 +1,0 @@
-
-{% for i in range(tasks) %}
-
-- name: null_{{ i }}
-  cmd: null
-  stdout: /dev/null
-
-{% endfor %}
