--- test/templates/should_fail/dependency_failures.jst
@@ -1,12 +1,0 @@
-
-- name: test_failures
-  stdout: /dev/null
-  cmd: |
-    echo "This task should fail"
-    exit 1
-
-
-- name: test_dependency_failures
-  after: test_failures
-  stdout: /dev/null
-  cmd: echo This command should fail due to dependency
