--- test/templates/should_fail/dependency_failures4.jst
@@ -1,16 +1,0 @@
-- name: task1
-  cmd: |
-    echo this should  run
-
-- name: task2
-  after: task1
-  exec: |
-    raise ValueError('foo')
-  cmd: echo this should not run
-
-{% for i in range(3, 100) %}
-- name: task{{ i }}
-  after: task{{ i - 1 }}
-  cmd: echo this should not run either
-
-{% endfor %}
