--- test/templates/should_fail/no_matching_task_name.jst
@@ -1,4 +1,0 @@
-
-- name: foo
-  after: bar
-  cmd: "true"
