--- test/templates/should_fail/no_matching_task_output.jst
@@ -1,4 +1,0 @@
-
-- name: foo
-  input: bar.txt
-  cmd: cat bar.txt
