--- test/templates/should_pass/basic.jst
@@ -1,9 +1,0 @@
-
-- name: hello_world
-  stdout: /dev/null
-  cmd: echo Hello, World
-
-- name: goodbye_world
-  stdout: /dev/null
-  after: hello_world
-  cmd: echo Goodbye, world
