--- test/templates/should_pass/dynamic.jst
@@ -1,33 +1,0 @@
-- name: start
-  stdout: /dev/null
-  cmd: |
-    set -ue
-
-    echo Dynamic workflow that will add tasks to the DAG while its running
-
-
-- name: add_tasks
-  stdout: /dev/null
-  after: start
-  exec: |
-    runner.workflow.new_task(
-        name='dynamic_task',
-        cmd='sleep 1'
-    )
-
-
-- name: after_adding_tasks
-  stdout: /dev/null
-  after: add_tasks
-  cmd: |
-    set -ue
-
-    echo The final workflow should have 5 tasks total
-
-
-- name: after_dynamic_task
-  stdout: /dev/null
-  after_re: dynamic_task|after_adding_tasks
-  cmd: |
-    echo This should have run after \"dynamic_task\"
-
