--- test/templates/should_pass/exec_directives.jst
@@ -1,18 +1,0 @@
-
-- name: start
-  stdout: /dev/null
-  cmd: date
-
-
-- name: add_some_tasks
-  stdout: /dev/null
-  after: start
-  exec: |
-      for i in range(10):
-          runner.workflow.new_task(name='new_task_{}'.format(i), cmd='true')
-
-
-- name: finalize
-  stdout: /dev/null
-  after: add_some_tasks
-  cmd: date
