--- test/templates/should_pass/hello_world.jst
@@ -1,3 +1,0 @@
-- name: hello_world
-  stdout: /dev/null
-  cmd: echo Hello, world!
