--- test/templates/should_pass/inheritance.jst
@@ -1,5 +1,0 @@
-{% include "hello_world.jst" %}
-
-- name: get_hostname
-  cmd: hostname
-  stdout: /dev/null
