--- test/templates/should_pass/input_output.jst
@@ -1,11 +1,0 @@
-
-- cmd: hostname
-  stdout: /dev/null
-  input: file.txt
-
-- cmd: echo make file
-  stdout: /dev/null
-  output: file.txt
-
-- cmd: echo yay
-  stdout: /dev/null
