--- test/templates/should_pass/mapping.jst
@@ -1,10 +1,0 @@
-name: myworkflow
-properties: can be added here
-tasks:
-
-- name: task1
-  cmd: sleep 1
-
-- name: task2
-  after: task1
-  cmd: sleep 1
