--- test/templates/should_pass/null.jst
@@ -1,8 +1,0 @@
-
-{% for i in range(tasks) %}
-
-- name: stress_null_{{ i }}
-  cmd: null
-  stdout: /dev/null
-
-{% endfor %}
