--- test/templates/should_pass/retry.jst
@@ -1,12 +1,0 @@
-- name: fails_once
-  retry: 1
-  stdout: /dev/null
-  cmd: |
-    if [ ! -f foo.txt ]; then
-      echo "File not found!"
-      touch foo.txt
-      exit 1
-    else
-      echo "File was found!"
-    fi
-
