--- test/templates/should_pass/stdout.jst
@@ -1,3 +1,0 @@
-- name: test_pipe_stdout_back_to_runner
-  stdout: /dev/stdout
-  cmd: echo -e "\nThanks for using Jetstream!"
