--- test/templates/should_pass/test.jst
@@ -1,16 +1,0 @@
-
-- name: start
-  cmd: sleep 1
-
-- name: echo_hello
-  stdout: /dev/null
-  cmd: echo Hello, world.
-
-
-{% for i in range(5) %}
-- name: test_range_{{ loop.index }}
-  stdout: /dev/null
-  after: start
-  cmd: echo {{ loop.index }}
-{% endfor %}
-
