--- test/templates/should_pass/true.jst
@@ -1,9 +1,0 @@
-
-{% for i in range(tasks) %}
-
-- name: stress_true_{{ i }}
-  stdout: /dev/null
-  stdout: /dev/null
-  cmd: "true"
-
-{% endfor %}
