--- test/templates/sleep30.jst
@@ -1,7 +1,0 @@
-{% for i in range(tasks) %}
-
-- name: stress_sleep30_{{ i }}
-  cpus: 0
-  cmd: sleep 30
-
-{% endfor %}
