--- test/templates/true.jst
@@ -1,8 +1,0 @@
-
-{% for i in range(tasks) %}
-
-- name: stress_true_{{ i }}
-  stdout: /dev/null
-  cmd: "true"
-
-{% endfor %}
