--- test/templates/variables.yaml
@@ -1,3 +1,0 @@
-# This same config file is used to fill in variables for all templates
-# in the test_templates/files/ directory
-tasks: 10
